www/kd/ : cgi-uploads

+++ www/kd/cgi-uploads/ul.py
@@ -19,6 +19,10 @@ if "file" in form:
         file_name = file_item.filename
         file_path = './upload-py-' + file_name
 
+        if os.path.exists(file_path):
+            print("PYTHON : file exists\n");
+            exit (0)
+
         print("file : ", file_name)
         print("path : ", file_path)
         with open(file_path, 'wb') as file:
